jupyter notebook file on Newton Raphson method
Contains a brief description of the function and a demo execution of the function afterward. The function newton is better than the usual Newton-Raphson method execution because it does not require estimation of the derivative of the function fo execution, rather it uses a simple local forward difference approach to estimate local derivative, hence making it more generalized version.